Fix logic for waiting for login and reconnecting
