+++ tests/rcon/test_rcon.py
@@ -5,7 +5,7 @@ from typing import Union, Text, Tuple
 
 import pytest
 
-from carim_discord_bot.rcon import service, protocol
+from carim_discord_bot.rcon import service, protocol, registrar
 from carim_discord_bot import config
 from carim_discord_bot.rcon.protocol import FORMAT_PREFIX, PACKET_TYPE_FORMAT, SEQUENCE_NUMBER_FORMAT
 
@@ -14,6 +14,7 @@ logging.basicConfig(level=logging.DEBUG, format='%(asctime)s %(levelname)s %(nam
 
 class MockServerProtocol(asyncio.DatagramProtocol):
     def __init__(self):
+        self.login_success = True
         self.transport = None
 
     def connection_made(self, transport: asyncio.transports.BaseTransport) -> None:
@@ -23,10 +24,13 @@ class MockServerProtocol(asyncio.DatagramProtocol):
         packet = protocol.Packet.parse(data)
         if isinstance(packet.payload, protocol.Login):
             packet.payload =\
-                CustomPayload(protocol.LOGIN, struct.pack(FORMAT_PREFIX + SEQUENCE_NUMBER_FORMAT, protocol.SUCCESS))
+                CustomPayload(protocol.LOGIN, struct.pack(FORMAT_PREFIX + SEQUENCE_NUMBER_FORMAT,
+                                                          protocol.SUCCESS if self.login_success else 0x00))
+            self.transport.sendto(packet.generate(), addr)
         elif isinstance(packet.payload, protocol.Command):
             packet.payload = protocol.Command(packet.payload.sequence_number, command='random command data')
-        self.transport.sendto(packet.generate(), addr)
+            if self.login_success:
+                self.transport.sendto(packet.generate(), addr)
 
 
 class CustomPayload(protocol.Payload):
@@ -41,36 +45,29 @@ class CustomPayload(protocol.Payload):
         raise NotImplementedError
 
 
-def loop_exception_handler(loop, context):
-    return
-    loop.default_exception_handler(context)
-
-
+@pytest.mark.timeout(5)
 @pytest.mark.asyncio
 async def test_login_success_before_other_commands(event_loop: asyncio.BaseEventLoop):
-    event_loop.set_exception_handler(loop_exception_handler)
     config.set(config.Config.build_from_dict({
         'token': '',
         'rcon_ip': '127.0.0.1',
-        'rcon_port': 2302,
+        'rcon_port': 42302,
         'rcon_password': 'password',
         'rcon_keep_alive_interval': .2,
     }))
     future_queue = asyncio.Queue()
     event_queue = asyncio.Queue()
     chat_queue = asyncio.Queue()
+    registrar.default_timeout = 1
+    mock_protocol = MockServerProtocol()
     server_t, server_p = await asyncio.get_running_loop().create_datagram_endpoint(
-        lambda: MockServerProtocol(), local_addr=(config.get().ip, config.get().port))
+        lambda: mock_protocol, local_addr=(config.get().ip, config.get().port))
 
     future = event_loop.create_future()
     await future_queue.put((future, 'players'))
     await service.start(future_queue, event_queue, chat_queue)
-
-    with pytest.raises(asyncio.CancelledError):
-        result = await future
-
-    async with service.process_command:
-        pass
+    result = await future
+    assert result == 'random command data'
 
     future = event_loop.create_future()
     await future_queue.put((future, 'players'))
